+++ src/molecules/collapse/Collapse.vue
@@ -4,6 +4,7 @@
 import AIcon from '../../atoms/icon1/Icon.vue'
 import AIconExpandMore from '../../atoms/icon1/templates/IconExpandMore.vue'
 import AButton from '../../atoms/button/Button.vue'
+import ATransitionExpand from '../../atoms/transition-expand/TransitionExpand.vue'
 
 import ACollapse from './Collapse.js'
 
@@ -11,8 +12,9 @@ export default {
   name: 'AlpacaCollapse',
   components: {
     AIcon,
-    AIconExpandMore,
+    ATransitionExpand
     AButton
+    AIconExpandMore,
   },
   mixins: [ACollapse]
 }