+++ src/molecules/select/Select.vue
@@ -4,9 +4,13 @@
 import ASelect from './Select.js'
 
 import ALabel from '../../atoms/label/Label.vue'
+import AIcon from '../../atoms/icon/Icon.vue'
 
 export default {
-  components: { ALabel },
+  components: {
+    ALabel,
+    AIcon
+  },
   mixins: [ASelect]
 }
 </script>