+++ src/organisms/footer/Footer.vue
@@ -1,14 +1,15 @@
 <template src="./Footer.html" />
 
 <script>
+import AContainer from '../../atoms/container/Container.vue'
 import AIcon from '../../atoms/icon/Icon.vue'
 import ALink from '../../atoms/link/Link.vue'
 import AButton from '../../atoms/button/Button.vue'
 import AList from '../../atoms/list/List.vue'
 import AListItem from '../../atoms/list-item/ListItem.vue'
 import ACookieMessage from '../../molecules/cookie-message/CookieMessage.vue'
-import ADropdownList from '../../molecules/collapse/Collapse.vue'
-import ANewsletter from '../../03-modules/newsletter/Newsletter.vue'
+import AFooterMenu from '../../molecules/footer-menu/FooterMenu.vue'
+import ANewsletter from '../../molecules/newsletter/Newsletter.vue'
 
 import AFooter from './Footer.js'
 
@@ -16,12 +17,13 @@ export default {
   name: 'AlpacaFooter',
   components: {
     AButton,
+    AContainer,
     ALink,
     AIcon,
     AList,
     AListItem,
     ACookieMessage,
-    ADropdownList,
+    AFooterMenu,
     ANewsletter
   },
   mixins: [AFooter]